+++ api-scanner/src/test/java/net/corda/plugins/BasicClassTest.java
@@ -23,6 +23,6 @@ public class BasicClassTest {
             "public class net.corda.example.BasicClass extends java.lang.Object\n" +
             "  public <init>(String)\n" +
             "  public String getName()\n" +
-            "##", CopyUtils.toString(testProject.getApi()));
+            "##", testProject.getApiText());
     }
 }