+++ api-scanner/src/test/java/net/corda/plugins/BasicInterfaceTest.java
@@ -22,6 +22,6 @@ public class BasicInterfaceTest {
         assertEquals(
             "public interface net.corda.example.BasicInterface\n" +
             "  public abstract java.math.BigInteger getBigNumber()\n" +
-            "##", CopyUtils.toString(testProject.getApi()));
+            "##", testProject.getApiText());
     }
 }