+++ api-scanner/src/test/java/net/corda/plugins/ExtendedClassTest.java
@@ -1,6 +1,6 @@
 package net.corda.plugins;
 
-import org.junit.Rule;
+import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.rules.RuleChain;
 import org.junit.rules.TemporaryFolder;
@@ -8,20 +8,29 @@ import org.junit.rules.TestRule;
 
 import java.io.IOException;
 
-import static org.junit.Assert.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class ExtendedClassTest {
-    private final TemporaryFolder testProjectDir = new TemporaryFolder();
-    private final GradleProject testProject = new GradleProject(testProjectDir, "extended-class");
+    private static final TemporaryFolder testProjectDir = new TemporaryFolder();
+    private static final GradleProject testProject = new GradleProject(testProjectDir, "extended-class");
 
-    @Rule
-    public TestRule rules = RuleChain.outerRule(testProjectDir).around(testProject);
+    @ClassRule
+    public static final TestRule rules = RuleChain.outerRule(testProjectDir).around(testProject);
 
     @Test
     public void testExtendedClass() throws IOException {
-        assertEquals(
-            "public class net.corda.example.ExtendedClass extends java.io.FilterInputStream\n" +
-            "  public <init>(java.io.InputStream)\n" +
-            "##", CopyUtils.toString(testProject.getApi()));
+        assertThat(testProject.getApiLines()).containsSequence(
+            "public class net.corda.example.ExtendedClass extends java.io.FilterInputStream",
+            "  public <init>(java.io.InputStream)",
+            "##");
+    }
+
+    @Test
+    public void testImplementingClass() throws IOException {
+        assertThat(testProject.getApiLines()).containsSequence(
+            "public class net.corda.example.ImplementingClass extends java.lang.Object implements java.io.Closeable",
+            "  public <init>()",
+            "  public void close()",
+            "##");
     }
 }