+++ api-scanner/src/test/java/net/corda/plugins/GradleProject.java
@@ -8,8 +8,16 @@ import org.junit.rules.TestRule;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
 
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
+import static java.util.Collections.emptyList;
+import static java.util.stream.Collectors.joining;
 import static net.corda.plugins.CopyUtils.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.gradle.testkit.runner.TaskOutcome.*;
@@ -19,6 +27,8 @@ import static org.junit.Assert.*;
  * JUnit rule to execute the scanApi Gradle task. This rule should be chained with TemporaryFolder.
  */
 public class GradleProject implements TestRule {
+    private static final String testGradleUserHome = System.getProperty("test.gradle.user.home", "");
+
     private final TemporaryFolder projectDir;
     private final String name;
 
@@ -28,10 +38,20 @@ public class GradleProject implements TestRule {
     public GradleProject(TemporaryFolder projectDir, String name) {
         this.projectDir = projectDir;
         this.name = name;
+        this.output = "";
     }
 
     public Path getApi() {
         return api;
+    }
+
+    public List<String> getApiLines() throws IOException {
+        // Files.readAllLines() uses UTF-8 by default.
+        return (api == null) ? emptyList() : Files.readAllLines(api);
+    }
+
+    public String getApiText() throws IOException {
+        return getApiLines().stream().collect(joining("\n"));
     }
 
     public String getOutput() {
@@ -63,5 +83,19 @@ public class GradleProject implements TestRule {
                 base.evaluate();
             }
         };
+    }
+
+    public static Path pathOf(TemporaryFolder folder, String... elements) {
+        return Paths.get(folder.getRoot().getAbsolutePath(), elements);
+    }
+
+    public static List<String> getGradleArgsForTasks(String... taskNames) {
+        List<String> args = new ArrayList<>(taskNames.length + 3);
+        Collections.addAll(args, taskNames);
+        args.add("--info");
+        if (!testGradleUserHome.isEmpty()) {
+            Collections.addAll(args,"-g", testGradleUserHome);
+        }
+        return args;
     }
 }