+++ api-scanner/src/test/java/net/corda/plugins/InternalMethodTest.java
@@ -22,6 +22,6 @@ public class InternalMethodTest {
         assertEquals(
             "public class net.corda.example.WithInternalMethod extends java.lang.Object\n" +
             "  public <init>()\n" +
-            "##", CopyUtils.toString(testProject.getApi()));
+            "##", testProject.getApiText());
     }
 }