+++ api-scanner/src/test/java/net/corda/plugins/KotlinAnnotationsTest.java
@@ -1,22 +1,21 @@
 package net.corda.plugins;
 
-import org.junit.Rule;
+import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.rules.RuleChain;
 import org.junit.rules.TemporaryFolder;
 import org.junit.rules.TestRule;
 
 import java.io.IOException;
-import java.nio.file.Files;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class KotlinAnnotationsTest {
-    private final TemporaryFolder testProjectDir = new TemporaryFolder();
-    private final GradleProject testProject = new GradleProject(testProjectDir, "kotlin-annotations");
+    private static final TemporaryFolder testProjectDir = new TemporaryFolder();
+    private static final GradleProject testProject = new GradleProject(testProjectDir, "kotlin-annotations");
 
-    @Rule
-    public TestRule rules = RuleChain.outerRule(testProjectDir).around(testProject);
+    @ClassRule
+    public static final TestRule rules = RuleChain.outerRule(testProjectDir).around(testProject);
 
     private static final String[] expectedClassWithDeprecatedFunctions = {
         "public final class net.corda.example.HasDeprecatedFunctions extends java.lang.Object",
@@ -63,12 +62,37 @@ public class KotlinAnnotationsTest {
     };
 
     @Test
-    public void testKotlinAnnotations() throws IOException {
-        assertThat(Files.readAllLines(testProject.getApi()))
-            .containsSequence(expectedClassWithDeprecatedFunctions)
-            .containsSequence(expectedClassWithJvmField)
+    public void testDeprecatedAnnotation() throws IOException {
+        assertThat(testProject.getApiLines())
+            .containsSequence(expectedClassWithDeprecatedFunctions);
+    }
+
+    @Test
+    public void testJvmFieldAnnotation() throws IOException {
+        assertThat(testProject.getApiLines())
+            .containsSequence(expectedClassWithJvmField);
+    }
+
+    @Test
+    public void testJvmStaticAnnotation() throws IOException {
+        assertThat(testProject.getApiLines())
             .containsSequence(expectedClassWithJvmStaticFunction)
-            .containsSequence(expectedClassWithJvmStaticFunctionCompanion)
+            .containsSequence(expectedClassWithJvmStaticFunctionCompanion);
+    }
+
+    @Test
+    public void testJvmOverloadedAnnotation() throws IOException {
+        assertThat(testProject.getApiLines())
             .containsSequence(expectedClassWithOverloadedConstructor);
+    }
+
+    @Test
+    public void testJvmDefaultAnnotation() throws IOException {
+        assertThat(testProject.getApiLines())
+            .containsSequence(
+                "public interface net.corda.example.HasDefaultMethod",
+                 "  public void doSomething(String)",
+                 "##"
+            );
     }
 }