+++ api-scanner/src/test/java/net/corda/plugins/MethodWithInternalAnnotationTest.java
@@ -27,6 +27,6 @@ public class MethodWithInternalAnnotationTest {
         assertEquals("public class net.corda.example.method.HasVisibleMethod extends java.lang.Object\n" +
             "  public <init>()\n" +
             "  public void hasInvisibleAnnotations()\n" +
-            "##", CopyUtils.toString(testProject.getApi()));
+            "##", testProject.getApiText());
     }
 }