+++ api-scanner/src/test/resources/kotlin-annotations/kotlin/net/corda/example/HasDeprecatedFunctions.kt
@@ -1,3 +1,4 @@
+@file:Suppress("UNUSED")
 package net.corda.example
 
 class HasDeprecatedFunctions @Deprecated("Dont use this anymore!") constructor () {